+++ src/node/internal/public_process.ts
@@ -13,7 +13,6 @@ import {
   ERR_INVALID_ARG_TYPE,
   ERR_INVALID_ARG_VALUE,
   ERR_OUT_OF_RANGE,
-  ERR_UNSUPPORTED_OPERATION,
 } from 'node-internal:internal_errors';
 import processImpl from 'node-internal:process';
 import { Buffer } from 'node-internal:internal_buffer';
@@ -34,9 +33,6 @@ import { validateString } from 'node-internal:validators';
 import type { Readable } from 'node-internal:streams_readable';
 
 export { platform, nextTick, emitWarning, env, features };
-
-const workerdExperimental = !!Cloudflare.compatibilityFlags['experimental'];
-const nodeJsCompat = !!Cloudflare.compatibilityFlags['nodejs_compat'];
 
 // For stdin, we emulate `node foo.js < /dev/null`
 export const stdin = new ReadStream(null, {
@@ -251,9 +247,6 @@ export function uptime(): number {
 }
 
 export function loadEnvFile(path: string | URL | Buffer = '.env'): void {
-  if (!workerdExperimental || !nodeJsCompat) {
-    throw new ERR_UNSUPPORTED_OPERATION();
-  }
   const { readFileSync } = process.getBuiltinModule('fs') as typeof NodeFS;
   const parsed = parseEnv(
     readFileSync(path instanceof URL ? path : path.toString(), 'utf8')
