+++ src/workerd/api/node/tests/process-nodejs-test.wd-test
@@ -8,12 +8,11 @@ const unitTests :Workerd.Config = (
           (name = "worker", esModule = embed "process-nodejs-test.js"),
           (name = "mod", esModule = "export const { FOO } = process.env;")
         ],
-        compatibilityDate = "2024-12-28",
+        compatibilityDate = "2025-05-01",
         compatibilityFlags = [
-          "nodejs_compat_populate_process_env",
+          "nodejs_compat",
           "enable_nodejs_process_v2",
-          "enable_nodejs_fs_module",
-          "experimental"
+          "enable_nodejs_fs_module"
         ],
         bindings = [
           (name = "FOO", text = "BAR"),
