+++ src/workerd/api/node/tests/process-stdio-fs-nodejs-test.expected_stderr
@@ -1,0 +1,1 @@
+Direct writeSync to stderr
