+++ src/workerd/api/node/tests/process-stdio-fs-nodejs-test.wd-test
@@ -2,14 +2,15 @@ using Workerd = import "/workerd/workerd.capnp";
 
 const unitTests :Workerd.Config = (
   services = [
-    ( name = "process-stdio-test",
+    ( name = "process-stdio-fs-test",
       worker = (
         modules = [
-          (name = "worker", esModule = embed "process-stdio-nodejs-test.js")
+          (name = "worker", esModule = embed "process-stdio-fs-nodejs-test.js")
         ],
         compatibilityDate = "2025-05-01",
         compatibilityFlags = [
           "nodejs_compat",
+          "enable_nodejs_fs_module",
           "enable_nodejs_process_v2"
         ],
       )
