+++ src/workerd/api/node/tests/process-stdio-nodejs-test.expected_stderr
@@ -1,4 +1,3 @@
-Direct writeSync to stderr
 Error Line 1
 Error Line 2
 Error Line 3
