+++ src/workerd/api/node/tests/process-stdio-nodejs-test.js
@@ -1,5 +1,3 @@
-import { ReadStream } from 'node:fs';
-import { writeSync } from 'node:fs';
 import { Buffer } from 'node:buffer';
 import { Readable, Writable } from 'node:stream';
 import assert from 'node:assert';
@@ -10,10 +8,6 @@ export const processStdioPropertiesTest = {
     assert(
       process.stdin instanceof Readable,
       'stdin should be instance of Readable'
-    );
-    assert(
-      process.stdin instanceof ReadStream,
-      'stdin should be instance of ReadStream'
     );
 
     assert.strictEqual(process.stdout.fd, 1, 'stdout should have fd 1');
@@ -181,43 +175,6 @@ export const processStdinTest = {
 
     assert.strictEqual(dataReceived, false);
     assert.strictEqual(endReceived, true);
-  },
-};
-
-export const fdBasedOperationsTest = {
-  test() {
-    const message = 'Direct writeSync to stdout\n';
-    const buffer = Buffer.from(message);
-    const bytesWritten = writeSync(1, buffer);
-    assert.strictEqual(
-      bytesWritten,
-      buffer.length,
-      'writeSync should return correct byte count'
-    );
-
-    const errorMessage = 'Direct writeSync to stderr\n';
-    const errorBuffer = Buffer.from(errorMessage);
-    const errorBytesWritten = writeSync(2, errorBuffer);
-    assert.strictEqual(
-      errorBytesWritten,
-      errorBuffer.length,
-      'writeSync to stderr should return correct byte count'
-    );
-  },
-};
-
-export const largeWriteTruncationTest = {
-  test() {
-    const largeBuffer = Buffer.alloc(4001 * 5, 'A'.repeat(4000) + '\n');
-    largeBuffer[0] = 66;
-    largeBuffer[4001 * 5 - 1] = 67;
-    const bytesWritten = writeSync(1, largeBuffer);
-
-    assert.strictEqual(
-      bytesWritten,
-      16 * 1024,
-      'Direct fd write should be truncated to 16KiB'
-    );
   },
 };
